+++ src/spells/mesh.h
@@ -94,6 +94,7 @@ public:
 	}
 
 	QModelIndex cast( NifModel * nif, const QModelIndex & index ) override;
+	static void updateBSTriShapeDataSize( NifModel * nif, const QModelIndex & index );
 };
 
 #endif